Added input validation for start.gg link
